refactor: rename server/ to engine/ + performance optimizations
## Structural Changes
- Renamed server/ folder to engine/ for clearer naming
- Updated all references in Dockerfile, CI, Makefile, docs

## Performance Optimizations
- Changed Job.data from Value to Arc<Value> for zero-copy cloning
- Added has_event_listeners() for lazy event allocation
- Optimized ack_batch with queue name caching
- Single lock scope in fail() for reduced contention

## Docker
- Commented out postgres, grafana, prometheus in docker-compose.yml
- Simplified for standalone flashq development

## Test Fixes
- Fixed test_kv_value_too_large (now uses 11MB string)
- Fixed clippy warnings (unnecessary cast, PI constant)

+++ engine/src/queue/tests/core.rs
@@ -144,7 +144,7 @@ async fn test_push_batch_large() {
     let inputs: Vec<_> = (0..100)
         .map(|i| crate::protocol::JobInput {
             data: json!({"i": i}),
-            priority: i as i32,
+            priority: i,
             delay: None,
             ttl: None,
             timeout: None,
